Feature: Supports stock buy and sell units in floating point. (#3193411) Feature: Supports stock merging and splitting.

+++ src/org/yccheok/jstock/gui/SellPortfolioTreeTableModel.java
@@ -48,7 +48,7 @@ public class SellPortfolioTreeTableModel extends AbstractPortfolioTreeTableModel
     private static final Class[] cTypes = {
         TreeTableModel.class,
         org.yccheok.jstock.engine.SimpleDate.class,
-        Integer.class,
+        Double.class,
         Double.class,
         Double.class,
         Double.class,
